# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Resource Hiring API
  summary: Production API contract for authentication and employee administration.
  version: 1.0.0
  description: |
    Smart Resource Hiring exposes REST APIs for user authentication and
    ADMIN-only employee account management.

    All request and response bodies use JSON unless an endpoint explicitly
    documents a different media type. Protected endpoints require a JWT bearer
    token in the `Authorization` header.
  contact:
    name: SRH API Team
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Authentication
    description: Login and session termination endpoints.
  - name: Employees
    description: ADMIN-only employee account management endpoints.
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      operationId: login
      summary: Authenticate user
      description: |
        Authenticates a registered employee by email and password.

        On success, the API returns a JWT access token containing the employee
        email and role. Use this token as `Authorization: Bearer <token>` when
        calling protected APIs.
      security: []
      requestBody:
        required: true
        description: Registered email and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              adminLogin:
                $ref: '#/components/examples/LoginRequestAdmin'
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  $ref: '#/components/examples/LoginResponseAdmin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/auth/logout:
    post:
      tags:
        - Authentication
      operationId: logout
      summary: End client session
      description: |
        Validates that the request includes a bearer token header and returns a
        successful logout response.

        JWTs are stateless in the current backend implementation, so logout is
        completed by the client discarding the token. The server does not keep a
        token revocation list.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout request accepted.
          content:
            text/plain:
              schema:
                type: string
                example: Logged out successfully
        '400':
          description: Missing or malformed Authorization header.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTokenFormat:
                  $ref: '#/components/examples/InvalidTokenFormatError'

  /api/employees:
    get:
      tags:
        - Employees
      operationId: listEmployees
      summary: List employees
      description: |
        Returns all employee accounts in the system.

        Requires a valid JWT for a user with the `ADMIN` role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Employees returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeResponse'
              examples:
                employees:
                  $ref: '#/components/examples/EmployeeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Employees
      operationId: createEmployee
      summary: Create employee
      description: |
        Creates a new employee account and stores the supplied password as a
        BCrypt hash.

        Requires a valid JWT for a user with the `ADMIN` role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Employee profile and temporary login password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
            examples:
              employee:
                $ref: '#/components/examples/EmployeeCreateRequest'
      responses:
        '201':
          description: Employee created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
              examples:
                created:
                  $ref: '#/components/examples/EmployeeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/employees/{id}:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'

    get:
      tags:
        - Employees
      operationId: getEmployeeById
      summary: Get employee by ID
      description: |
        Returns a single employee account by database ID.

        Requires a valid JWT for a user with the `ADMIN` role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Employee returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
              examples:
                employee:
                  $ref: '#/components/examples/EmployeeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Employees
      operationId: updateEmployee
      summary: Update employee
      description: |
        Updates the supplied employee fields.

        Omitted fields are left unchanged. If a password is supplied, it must be
        valid and will be stored as a BCrypt hash. Requires a valid JWT for a
        user with the `ADMIN` role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Fields to update for the employee.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
            examples:
              roleUpdate:
                $ref: '#/components/examples/EmployeeUpdateRequest'
      responses:
        '200':
          description: Employee updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
              examples:
                updated:
                  $ref: '#/components/examples/EmployeeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags:
        - Employees
      operationId: deleteEmployee
      summary: Delete employee
      description: |
        Deletes an employee account by database ID.

        Requires a valid JWT for a user with the `ADMIN` role.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Employee deleted successfully. The response body is empty.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send the access token from `POST /api/auth/login` as
        `Authorization: Bearer <token>`.

  parameters:
    EmployeeId:
      name: id
      in: path
      required: true
      description: Unique database ID of the employee.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1

  responses:
    BadRequest:
      description: The request body, parameters, or headers are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              $ref: '#/components/examples/ValidationError'
    Unauthorized:
      description: Authentication failed or the JWT is missing, expired, or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              $ref: '#/components/examples/UnauthorizedError'
    Forbidden:
      description: The authenticated user does not have the required ADMIN role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              $ref: '#/components/examples/ForbiddenError'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              $ref: '#/components/examples/EmployeeNotFoundError'
    Conflict:
      description: The request conflicts with an existing employee record.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            duplicateEmail:
              $ref: '#/components/examples/DuplicateEmailError'

  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Registered employee email address.
          example: admin@example.com
        password:
          type: string
          format: password
          description: Employee password.
          minLength: 1
          example: admin123

    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - token
        - role
        - email
      properties:
        token:
          type: string
          description: JWT access token used for protected API calls.
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbkBleGFtcGxlLmNvbSJ9.signature
        role:
          $ref: '#/components/schemas/Role'
        email:
          type: string
          format: email
          description: Authenticated employee email address.
          example: admin@example.com

    EmployeeRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          description: Employee display name.
          minLength: 1
          example: Aarav Sharma
        email:
          type: string
          format: email
          description: Unique employee email address.
          example: aarav.sharma@example.com
        password:
          type: string
          format: password
          description: Temporary login password. The API stores only a BCrypt hash.
          minLength: 6
          maxLength: 72
          example: employee123
        role:
          $ref: '#/components/schemas/Role'

    EmployeeUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      description: |
        Partial employee update payload. Omit a field to keep its current value.
      properties:
        name:
          type: string
          description: Replacement employee display name.
          minLength: 1
          example: Aarav Sharma
        email:
          type: string
          format: email
          description: Replacement unique employee email address.
          example: aarav.sharma@example.com
        password:
          type: string
          format: password
          description: Optional replacement password. The API stores only a BCrypt hash.
          minLength: 6
          maxLength: 72
          example: newPassword123
        role:
          $ref: '#/components/schemas/Role'

    EmployeeResponse:
      type: object
      additionalProperties: false
      description: Employee details returned by the API. Passwords are never returned.
      required:
        - id
        - name
        - email
        - role
      properties:
        id:
          type: integer
          format: int64
          description: Unique database ID of the employee.
          example: 1
        name:
          type: string
          description: Employee display name.
          example: Aarav Sharma
        email:
          type: string
          format: email
          description: Employee email address.
          example: aarav.sharma@example.com
        role:
          $ref: '#/components/schemas/Role'

    Role:
      type: string
      description: Employee authorization role.
      enum:
        - ADMIN
        - EMPLOYEE
        - OPERATOR
        - PROJECT_ADMIN
      example: EMPLOYEE

    ErrorResponse:
      type: object
      additionalProperties: false
      description: Standard API error response returned by application exception handlers.
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when the error response was generated.
          example: "2026-06-08T07:30:00Z"
        status:
          type: integer
          format: int32
          description: HTTP status code.
          example: 401
        error:
          type: string
          description: HTTP status reason phrase.
          example: Unauthorized
        message:
          type: string
          description: Human-readable error detail.
          example: Invalid email or password
        path:
          type: string
          description: Request path that produced the error.
          example: /api/auth/login

  examples:
    LoginRequestAdmin:
      summary: Admin login
      value:
        email: admin@example.com
        password: admin123

    LoginResponseAdmin:
      summary: Authenticated admin
      value:
        token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbkBleGFtcGxlLmNvbSJ9.signature
        role: ADMIN
        email: admin@example.com

    EmployeeCreateRequest:
      summary: Create employee account
      value:
        name: Aarav Sharma
        email: aarav.sharma@example.com
        password: employee123
        role: EMPLOYEE

    EmployeeUpdateRequest:
      summary: Promote employee role
      value:
        role: PROJECT_ADMIN

    EmployeeResponse:
      summary: Employee details
      value:
        id: 1
        name: Aarav Sharma
        email: aarav.sharma@example.com
        role: EMPLOYEE

    EmployeeListResponse:
      summary: Employee list
      value:
        - id: 1
          name: Aarav Sharma
          email: aarav.sharma@example.com
          role: EMPLOYEE
        - id: 2
          name: Kavya Nair
          email: kavya.nair@example.com
          role: PROJECT_ADMIN

    ValidationError:
      summary: Validation failure
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 400
        error: Bad Request
        message: "email: must be a well-formed email address, password: size must be between 6 and 72"
        path: /api/employees

    UnauthorizedError:
      summary: Invalid credentials or token
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 401
        error: Unauthorized
        message: Invalid email or password
        path: /api/auth/login

    ForbiddenError:
      summary: Missing ADMIN role
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 403
        error: Forbidden
        message: Access Denied
        path: /api/employees

    EmployeeNotFoundError:
      summary: Employee not found
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 404
        error: Not Found
        message: Employee not found
        path: /api/employees/999

    DuplicateEmailError:
      summary: Duplicate employee email
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 409
        error: Conflict
        message: Employee email already exists
        path: /api/employees

    InvalidTokenFormatError:
      summary: Invalid bearer token format
      value:
        timestamp: "2026-06-08T07:30:00Z"
        status: 400
        error: Bad Request
        message: Invalid token format
        path: /api/auth/logout